display polls and posts on tag page
